requirements.txt
# Core Dependencies
openai==1.54.3
pydantic==2.9.2
langfuse==2.52.2
python-dotenv==1.0.1

# Image/PDF Processing
pillow==10.4.0
PyMuPDF==1.26.7

# Web UI
streamlit==1.52.1

# Testing
pytest==8.3.3
pytest-mock==3.14.0
